# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhaleWatch API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/investors/{slug}/portfolio": {
      "get": {
        "tags": [
          "investors"
        ],
        "summary": "One investor's positions in one period",
        "description": "The portfolio, paginated, with each position's change since the filer's\nprevious published period.\n\n``meta.caveats`` lists what ``audit-amendments`` would flag about the\nperiod, so a client sees it on every page.",
        "operationId": "getInvestorPortfolio",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The investor's slug, as `slug` on its row of `GET /v1/investors`.",
              "examples": [
                "berkshire-hathaway"
              ],
              "title": "Slug"
            },
            "description": "The investor's slug, as `slug` on its row of `GET /v1/investors`."
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}([Qq]\\d|-\\d{2}-\\d{2})$",
                  "examples": [
                    "2026Q1",
                    "2026-03-31"
                  ]
                },
                {
                  "type": "null"
                }
              ],
              "description": "The quarter, as `2026Q1` or `2026-03-31`. Defaults to the latest one published for this investor, which `meta.period` states.",
              "title": "Period"
            },
            "description": "The quarter, as `2026Q1` or `2026-03-31`. Defaults to the latest one published for this investor, which `meta.period` states."
          },
          {
            "name": "include_options",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Add the period's option lines, one row per security and `put_call`. Their value is the notional of the underlying, so they have no weight, and weights stay shares of the common-stock portfolio.",
              "default": false,
              "title": "Include Options"
            },
            "description": "Add the period's option lines, one row per security and `put_call`. Their value is the notional of the underlying, so they have no weight, and weights stay shares of the common-stock portfolio."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PortfolioSort",
              "description": "`weight` (the same order as `value`, with option lines after stock), `value`, `shares`, or `change`: `shares_delta_pct`, with new positions as the largest increase and option lines last.",
              "default": "weight"
            },
            "description": "`weight` (the same order as `value`, with option lines after stock), `value`, `shares`, or `change`: `shares_delta_pct`, with new positions as the largest increase and option lines last."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Largest first by default.",
              "default": "desc"
            },
            "description": "Largest first by default."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Rows per page. Above 200 is refused, not cut down.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Rows per page. Above 200 is refused, not cut down."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`page.next_cursor` from the previous page. Omit for the first page. Keep the other parameters the same while paging.",
              "title": "Cursor"
            },
            "description": "`page.next_cursor` from the previous page. Omit for the first page. Keep the other parameters the same while paging."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioPositionEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No investor with that slug, or nothing published for the period.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "400": {
            "description": "`cursor` is malformed, or came from a different listing or sort order.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "A parameter is missing or malformed. `errors` says which.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/investors/{slug}/history": {
      "get": {
        "tags": [
          "investors"
        ],
        "summary": "One investor's portfolio value, positions and concentration, quarter by quarter",
        "description": "Every quarter from the first published to the latest, oldest first.\n\nA quarter in between with nothing published is a row of nulls, so a chart\nof it shows the gap rather than drawing across it. Empty for a filer with\nnothing published. As of the views' last refresh. Not paginated.",
        "operationId": "getInvestorHistory",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The investor's slug, as `slug` on its row of `GET /v1/investors`.",
              "examples": [
                "berkshire-hathaway"
              ],
              "title": "Slug"
            },
            "description": "The investor's slug, as `slug` on its row of `GET /v1/investors`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPointEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No investor with that slug.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "A parameter is missing or malformed. `errors` says which.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PortfolioPositionEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PortfolioPosition"
            },
            "type": "array",
            "title": "Data",
            "description": "The rows, in the order the endpoint documents."
          },
          "meta": {
            "$ref": "#/components/schemas/Meta",
            "description": "What the rows describe, and when this answer was built."
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page"
              },
              {
                "type": "null"
              }
            ],
            "description": "How to get the next page. Null when the endpoint is not paged."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "PortfolioPositionEnvelope",
        "description": "One investor's positions in one period."
      },
      "PortfolioSort": {
        "type": "string",
        "enum": [
          "weight",
          "value",
          "shares",
          "change"
        ],
        "title": "PortfolioSort",
        "description": "What a portfolio is ordered by."
      },
      "Problem": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong, stable across releases. `invalid_cursor`: `cursor` is malformed, or belongs to a different listing or sort. `not_found`: no such thing, or nothing published for the period asked for. `validation_error`: a parameter is missing or malformed; `errors` says which.",
            "examples": [
              "not_found"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What went wrong, in a sentence for a person. Not a contract: do not parse it.",
            "examples": [
              "No investor 'berkshire'. GET /v1/investors lists every one."
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "Problem",
        "description": "Why the request was refused."
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "description": "Ascending or descending."
      },
...
      "ValidationProblem": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong, stable across releases. `invalid_cursor`: `cursor` is malformed, or belongs to a different listing or sort. `not_found`: no such thing, or nothing published for the period asked for. `validation_error`: a parameter is missing or malformed; `errors` says which.",
            "examples": [
              "not_found"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What went wrong, in a sentence for a person. Not a contract: do not parse it.",
            "examples": [
              "No investor 'berkshire'. GET /v1/investors lists every one."
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors",
            "description": "Every parameter that failed, each once. Never empty."
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail",
          "errors"
        ],
        "title": "ValidationProblem",
        "description": "The 422 body: a parameter is missing or malformed."
      }
    }
  }
}